+++ cmd/evm/testdata/25/exp.json
@@ -8,11 +8,11 @@
       "nonce": "0xad"
     },
     "0xc94f5374fce5edbc8e2a8697c15331677e6ebf0b": {
-      "balance": "0x854d00"
+      "balance": "0x1ec3000"
     }
   },
   "result": {
-    "stateRoot": "0x5139609e39f4d158a7d1ad1800908eb0349cea9b500a8273a6cf0a7e4392639b",
+    "stateRoot": "0xb056800260ffcf459b9acdfd9b213fce174bdfa53cfeaf505f0cfa9f411db860",
     "txRoot": "0x572690baf4898c2972446e56ecf0aa2a027c08a863927d2dce34472f0c5496fe",
     "receiptsRoot": "0x056b23fbba480696b65fe5a59b8f2148a1299103c4f57df839233af2cf4ca2d2",
     "logsHash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
