+++ cmd/evm/testdata/30/exp.json
@@ -10,10 +10,13 @@
     "0xd02d72e067e77158444ef2020ff2d325f929b363": {
       "balance": "0xfffffffb8390",
       "nonce": "0x3"
+    },
+    "0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba": {
+      "balance": "0x47c70"
     }
   },
   "result": {
-    "stateRoot": "0x3483124b6710486c9fb3e07975669c66924697c88cccdcc166af5e1218915c93",
+    "stateRoot": "0x6e7833d2d72d8a7074d89aac54e2ddcbe018bad9078e2a05db32b0bd1b3255fa",
     "txRoot": "0x013509c8563d41c0ae4bf38f2d6d19fc6512a1d0d6be045079c8c9f68bf45f9d",
     "receiptsRoot": "0x75308898d571eafb5cd8cde8278bf5b3d13c5f6ec074926de3bb895b519264e1",
     "logsHash": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
