+++ eth/tracers/internal/tracetest/testdata/prestate_tracer_with_diff_mode/create_failed.json
@@ -87,7 +87,7 @@
         "nonce": 1223933
       },
       "0x8f03f1a3f10c05e7cccf75c1fd10168e06659be7": {
-        "balance": "0x38079c19423e44b30e"
+        "balance": "0x3807bc244dbe20e89b"
       }
     }
   }
